+++ src/router/router.js
@@ -2,7 +2,7 @@ import { Routes, Route } from 'react-router-dom';
 import { ROUTE } from '../common/utils/constant';
 import Main from '../components/Main/Main';
 import NotFound from '../components/NotFound/NotFound';
-import Todo from '../components/Todo';
+import Todo from '../components/Todo/Todo';
 
 function Router() {
   return (
